added to the keyboard next and search buttons.
This is a very nice feature added using the xml attribute imeOptions. Notes: 1) the dedicated button only appears when the attribute singleLine is set to true. So we set this attribute also for the path field (will see if that's usable in the future...). of course the search button should be added a listener to activate the search option.