implemented key methods BookRepository.getBook and Book.getChildren.
getBook gets a book name (in Hebrew) and returns a corresponding Book object. The object is built from the corresponding book file (which resides under assets). A book holds the whole table of contents as described in the file. The method getChildren gets a path in the tree and returns the children. Next we should implement the dynamic auto completion based on these methods.