url: adding UrlProvider and TanachUrlProvider.
BH starting to implement the main feature of connecting the table of contents to the internet... UrlProvider is an helper class
that is used during the parsing of a book file and helps tracking the url of each leaf node. The class TanachUrlProvider is
a first concrete class the implements UrlProvider for the Tanach of Mechon Memre. Well done