changed soft keyboard policy: no always visible.
Why? don't like the idea that our clean screen will always be masked with a keyboard. Now the policy is:
not visible at startup (though I see that the keyboard is still opened at startup), and upon pressing
of the clear buttons the keyboard is opened (and the view is given focus. think it is better.