changed the theme due to problem with 2.4.
What happened? in our phone text was not visible in the text fields as well as the auto completion text. tried
several solutions and then found out that it is related to the theme. changed to a dark theme (Holo) and it solved the problem.
However now our background was black, so we set the background via the api to white. in the tablet we don't see any
difference (where in the phone we see a darker search button).